chore(refactor): remove slack users repository exception
The SlackUsersRepository populates itself from the Redis cache, if for
some reason the cache isn't populated it throws an exception. That has
proved to be problematic for the developers on their local machines as
they typically don't have the cache populated.

With this change, the repository would be empty if the cache is not
populated.