chore(tech-debt): add test for session controller: (#143)
The session controller handles logging of session, by
mentor or mentee and sending logged session to freckle.
This functionality is not impervious to breaks or bugs,
hence the need to test it. This commit test the
functionalities of this controller and goes ahead to
fix some of the bugs found in it. The bugs:
- someone else can log a session to a request they are
not part of
- someone else can approve a logged session of a request
they are not part of
were also fixed in this commit

[Finishes #149882416]