chore(session-and-skill): fix failing tests (#184)
'testRemoveFailure' for skills, failed sometimes
because faker was used to randomly seed the 'UserSkills'
table with test data. I modified the seeder such that
the skill being removed was always present on the table

'testLogSessionSuccess' in the session had an unnecessary
api call just to get the 'mentee_id' of a request,
the request data was not static and timezones for session date,
start_time and end_time were different. I removed
the extra api call and used static and consistent post data for
the test.

[Finishes #151147405]