feat(notifications): send guidelines on external mentorship to placed fellows (#206)
Initially, notification emails were only sent to success and they send the external mentorship emails.
This process is moved to lenken so that the system sends the guidelines on external mentorship.

I created a new email template that the placed fellows would receive that's rendered by the external mentorship class.
I then added the logic that sends the emails to the placed fellows by using the Mail inbuilt tool.

[Delivers # 151636859]