fix(request-report): fix checking of duration when fetching unmatched requests (#215)
* fix(requestModel): fix checking of duration when fetching unmatched requests

The logic for checking if duration parameter is set had a logical error,
as a result, it was causing the whole logic for fetching unmatched requests
to return the wrong value.

This fix rectifies the bug by refactoring that logic so that proper date
filters can be applied base on the duration value.

Two functions are implemented to separate the logic for fetching unmatched
requests before and after duration. ReportController is also updated to
reflect the update.

[Fixes #152072462]

* chore(refactor): separate duration from query params
The `buildQuery` function which take the params does not have
implementation for the `duration` key and therefore it makes sense to
revert it to being an independent argument.