feat(cancel-requests): allow admin close mentorship session (#207)
* feat(cancel-requests): admin can close mentorship session

This refactors the cancelRequest action of RequestsController to allow
the administrator to cancel a Request. It is a necessary fix since
without it an admin can not cancel a request from the front end

A change is made to the database to introduce a new table named
cancelled_requests . This table holds the data for all cancelled requests
and it is updated each time a request is cancelled

Update RequestControllerTest to include new test cases for updates
implemented on RequestsController

[Finishes #151641164]

* chore(refactor): clean-up implementation
Removed the error response for when a request cannot be cancelled. It's
easier to follow if the end of the logic is a success than when
successes are in-between and the ultimate end is a failure. Also a
failure like this should be something we look into not just respond to
the user.

Corrected the name of the migration that wasn't changed after the model
was renamed.