feat(notification): create command to send emails to fellows with inactive requests. (#214)
Currently no technique exists on getting the inactive requests. The implementation gathers
the inactive requests, extracts the mentor and mentee emails then sends them an email.

I created a command that checks
whether a session happened within the last 3 scheduled sessions. If none of the last three sessions
happened, the request is marked as inactive and an email sent to the mentor and mentee.

[Finishes #151798078]