feat(pool-filters): add functionality for users to filter requests pool (#228)
* feat(pool-filters): add functionality for users to filter requests pool

This implementation introduces filtering functionality for the
request pool. It is crucial that users are able to filter
requests as per their preferences.

I have refactored RequestController to process a new set of parameters
from the client. RequestController can now process filters for
multiple locations, multiple requests lengths, Ratings, and Recommended
Requests.

A new implementation is also added to SkillController to allow it to return
a unique list of skills which have atleast one request opened for them

[Finishes #152160424]

* chore(api-versioning): move new api implementation to v2

We've agreed that all new APIs or any functionality that requires modifying
the V1 endpoint should be put in V2.

I moved the getRequestSkills from V1 where it was placed to V2 and created
a new SkillController for it.
I removed the extra key "skills" cause it's unnecessary and ensured
the function returned an array of objects as against an array of arrays.

Note: the new endpoints still need to be tested.