feat(profile): create profile page (#229)
* feat(profile): create profile page

Currently, lenken V2 does not have a user profile feature.

This commit add a user profile feature to lenken V2, it ensure a lenken user can view and edit their profile.

[Finishes #152309363]

* chore(restructure): move user skill resource api to users controller
According to REST, we know the user skills resource belongs to the user,
the implementation for this should also be done in the User Controller.

I also broke up the test that was combined, tests should be independent
and check for one thing only.