feat(lenken-v2): add v2 user controller (#239)
* feat(lenken-v2): add V2 user controller

Lenken version 2 is currently being developed. This calls for some
changes to be made in the routes and controllers.

This PR adds functionality to the user controller and adds a users route.
When an api call is made to the route
http://lenken-dev.andela.com:3000/api/v2/users/*-userId*,the details
of the user is returned.

[Finishes #152536708]

* chore(refactor): cleanup codebase and rename variables for readability

* feat(lenken-v2): add V2 user controller

Lenken version 2 is currently being developed. This calls for some
changes to be made in the routes and controllers.

This PR adds functionality to the user controller and adds a users route.
When an api call is made to the route
http://lenken-dev.andela.com:3000/api/v2/users/*-userId*,the details
of the user is returned.

[Finishes #152536708]