ch(notification): improve tests for email notifications. (#247)
Currently the tests on email notifications simply test the number of recipients
that received the email regardless of the logic. It doesn't really test that the
emails have been sent to the right people.

I refactored the tests to get at least one intended recipient and affirm that
the logic actually send an email to that recipient. This also covers the fact
that emails were actually sent.

[Delivers #153114100]