chore(user-creation): create new user on getUserById
This basically ensures that when a user is gotten by ID through the
`users/{id}` endpoint, it creates the user if the user does not exist on
Lenken.

I also refactored the SlackUserRepositoryMock and the AISClientMock to
allow the tests pass since the user for the test doesn't have records in
those mock methods.

The endpoint for getting multiple users still need to be refactored as
it contains a lot of unnecessary that I ignored to move forward with
this task.

The unique key constraint on the slack_id for users has also been
removed because users such as "lenken-tech@andela.com" do not have slack
accounts and this will cause errors.

[Finishes #154411415]