chore(pending-pool): refactor formatRequestData method (#299)
* chore(pending-pool): refactor formatRequestData method

The formatRequestData method was modified to suit
the implementation of the getPendingPool feature.
This is not appropriate as the formatRequestData
is used by other methods in the RequestController.

This commit refactors the formatRequestData method
by taking out the added parameter, its use and
creating another private method to determine
the status of a pending request. This helps
to bring about a clear separation of concerns
among the methods.

* chore(refactor): cleanup code for readability

When appending we're not retuning new data, the concept is that we're
adding to the already existing one. Hence my changing the parameter to
be passed by reference so we only modify the existing array.

Also, the PHPDoc wasn't modified to suit the new function and we don't
name variables in snake case and with questions.


[Finishes #155850373]