chore(update-skills): inplement endpoing to enable and disable skills (#300)
Currently there is no endpoint to enable or disable skills

This commits creates an endpoint to enable and disable a skill.
To disable a skill I am  using laravel soft delete, so as  skill can
be soft deleted without being removed from the database.
Skills are enabled by restoring skills that are soft deleted.

 [Finishes #155768970]