chore(requests-table): remodel mentee-mentor-request relationship (#306)
* chore(requests-table): remodel mentee-mentor-request relationship
Lenken allows a user to make a mentorship request as a
mentee seeking a mentor.
On Lenken-v2 a user should be able to make a mentorship request either as a mentee
seeking a mentor, or a mentor seeking a mentee. This requires the
request table to be remodelled.

This commit removes the mentee_id and mentor_id fields in the request
table and adds created_by and request_type fields.
A new table called users_requests is created. The table has 2 entries
for a single request. For each entry the role of the user will be entered in
the role field. Role is either a mentor or mentee.

Two other tables are also created, role and request_type. These tables have
a defined collection of roles and request types respectively. The roles and
requests are referenced using their ids.
[Finishes #155582655]

* chore(requests-table): remodel mentee-mentor-request relationship

Lenken allows a user to make a mentorship request as a
mentee seeking a mentor.
On Lenken-v2 a user should be able to make a mentorship request either as a mentee
seeking a mentor, or a mentor seeking a mentee. This requires the
request table to be remodelled.

This commit removes the mentee_id and mentor_id fields in the request
table and adds created_by and request_type fields.
A new table called users_requests is created. The table has 2 entries
for a single request. For each entry the role of the user will be entered in
the role field. Role is either a mentor or mentee.

Two other tables are also created, role and request_type. These tables have
a defined collection of roles and request types respectively. The roles and
requests are referenced using their ids.

[Finishes #155582655]

* chore(refactor): include database params in the commands