chore(refactor): implement helpers to format request data
The formatRequestData function in the RequestController for V2 was
removed and and two helper functions, one to format a single request and
another to format multiple requests were implemented.

The seeking mentor/mentee naming for request types was also changed to
mentee/mentor requests for easy understanding.chore(refactor): implement helpers to format request data

The formatRequestData function in the RequestController for V2 was
removed and and two helper functions, one to format a single request and
another to format multiple requests were implemented.

The seeking mentor/mentee naming for request types was also changed to
mentee/mentor requests for easy understanding.