chore(request-table): make changes to accommodate new requests table (#315)
* chore(request-table): make changes to accommodate new requests table

We currently have failing tests that indicate some parts of the application
are broken following the changes made to the request table.

This commit refactors the code and ensures all tests are passing. This
ensures that the application is working as expected.

[Delivers #156302499]

* fix(database-query): fix query for requests by mentee id

* fix(migrations): modify migrations
This commit modifies the remodel_requests_table migration. It removes
the mentee_id foreign key relationship