chore(mentee/mentor rating): implement mentor/mentee rating (#320)
Currently, On the lenken app, only a mentee can rate
a mentor. Also, a logged session pending approval
cannot be edited and a logged session cannot be rejected.

This commit refactors the log session controller
to accomodate a mentor rating a mentee by implementing
a logic that checks the user to be rated before rating.

This commit  refactors the confirm session
controller to ensure when a mentor rates first,
the mentee can confirm the session.

[Finishes #156379943]