chore(requests): implement flag to get requests created by user (#349)
On the lenken server, there is currently no method
to get requests created by a single user. This
implies that a user cannot get all the requests
they have created.

This commit include a query parameter that returns
all the requests made by a user. This is implemented
by querying the 'all-request' route based on
the userId parameter.

[Finishes #156913801]