feat(freckle): add freckle integration on version two (#359)
Lenken version one has freckle integration implemented but this
is not implemented on version two.

This commit adds the freckle integration on version two. This commit
also adds the API landing page.

[Delivers #157389259]

* chore(freckle): refactor freckle integration

Cleanup the code as some extraneous things were brought in from V1.
This includes the email being sent out to the Mentor when a session
isn't logged.