chore(in-progress-endpoint): add endpoint to end in progress request (#355)
Currently, there exists no endpoint for ending an ongoing
mentorship request, this is needed to give ability to track
data related to the phase between an ongoing mentorship request
and when it is completed.

The commit provide an endpoint through which an in-progress
mentorship request can be ended.

[Finishes #157356958]

* chore(request-cancellation-model): update request-cancellation
model

The requestCancellation model has the same fields
as requestStatusUpdate except for the 'status_id' field.
The requestStatusUpdate model should be deprecated
so as to avoid redundant data in the database.

The requestCancellationReason model and migration
are updated to include the status_id
of the request and the requestStatusupdate
model and migration were removed.

* chore(refactor): clean up implementation for status update

The property and values expected from
the request has been changed to
accept the actual status ID instead
of the string for the status. This helps us get
leaner code.

The validators for cancellation and abandonment were move to the model and the cancellation
validation was modified to cater for the scenario
where the request to be cancelled has not been
matched.

The instance variable for status update was
placed in each test's function scope, to ensure proper context.

* fix(tests): fix failing tests

Some tests  in the application fail because of the
fragile assertions.

Also, some fail due to modifications made in the
requestController.

To fix the failing tests, changes were made in the
RequestController and the failing test files.